Spatial join service now accepts floating point values for `<maxDistance>` (#2093)
The `<maxDistance>` parameter of the spatial join service now accepts a floating point value and is also stored as a floating point value. In particular, the distance, which is measured in meters, can now be smaller than one meter. This follows the suggestion from <https://github.com/ad-freiburg/qlever/pull/1935#discussion_r2120272456>.

+++ src/parser/SpatialQuery.h
@@ -27,7 +27,7 @@ struct SpatialQuery : MagicServiceQuery {
 
   // The spatial join task definition: maximum distance and number of results.
   // One of both - or both - must be provided.
-  std::optional<size_t> maxDist_;
+  std::optional<double> maxDist_;
   std::optional<size_t> maxResults_;
 
   // Optional further argument: a variable to which the distance between spatial
